Initial restructuring of configuration loading mechanism
Microcosm has always had a `loader` abstraction for loading configuration
from a specific source; however, the overall mechanism of loading configuration
from defaults, merging in loaders, etc. has been inline, during object graph
creation. Now we have (the beginnnings) of an abstraction layer.
